chore: fix core/feature folder issues
