--- src-tauri/src/structs.rs
@@ -1,36 +1,0 @@
-#![allow(dead_code)]
-use serde::{Deserialize, Serialize};
-
-#[derive(Serialize, Deserialize)]
-pub struct Medication {
-    pub name: String,
-    pub brand: String,
-    pub dosage: f64,
-    pub frequency: f64,
-    pub supply: f64,
-    pub total_prescribed: f64,
-    pub first_added: f64,
-    pub last_taken: Option<f64>,
-    pub upcoming_dose: Option<f64>,
-    pub schedule: Option<String>,
-    pub measurement: String,
-    pub nurse_id: String,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct User {
-    pub id: String,
-    pub full_name: String,
-    pub type_of: String,
-    pub phone_number: Option<i64>,
-    pub email: Option<String>,
-}
-
-#[derive(Serialize, Deserialize)]
-pub struct MedicationLog {
-    pub timestamp: f64,
-    pub medication_name: String,
-    pub given_dose: f64,
-    pub measurement: String,
-    pub comment: Option<String>,
-}
